Support for root.vim and project.vim files
root.vim: File to designate a root project to build and the root of the
project file system.
project.vim: A smaller project that resides inside the root folder of
the project
also uploaded a small example of a root.vim. The same format is used for
the project.vim files.